Move out tests to compareUpdate helpers
