+++ test/services/AmendVariationServiceSpec.scala
@@ -498,144 +498,6 @@ class AmendVariationServiceSpec extends PlaySpec with MockitoSugar with ScalaFut
 
     "successfully compare and update api6 request with api5 data" when {
 
-      "responsible people start date changes" in {
-
-        val viewModel = DesConstants.SubscriptionViewStatusRP.copy(
-          responsiblePersons = DesConstants.SubscriptionViewStatusRP.responsiblePersons map {
-            rps => Seq(rps.head.copy(startDate = Some("1970-01-01")))
-          }
-        )
-        val request = DesConstants.updateAmendVariationRequestRP.copy(
-          responsiblePersons = DesConstants.updateAmendVariationRequestRP.responsiblePersons map {
-            rps => Seq(rps.tail.head)
-          }
-        )
-        val convertedRequest = DesConstants.amendStatusAmendVariationRP.copy(
-          responsiblePersons = DesConstants.amendStatusAmendVariationRP.responsiblePersons map {
-            rps => Seq(rps.tail.head.copy(dateChangeFlag = Some(true), extra = rps.tail.head.extra.copy(status = Some("Updated"))))
-          }
-          ,
-          aspOrTcsp = Some(DesConstants.testAspOrTcsp.copy(
-            supervisionDetails = Some(DesConstants.testSupervisionDetails.copy(
-              supervisorDetails = Some(DesConstants.testSupervisorDetails.copy(
-                dateChangeFlag = Some(false)
-              ))
-            ))
-          )))
-        when {
-          TestAmendVariationService.viewDesConnector.view(eqTo(amlsRegistrationNumber))(any())
-        } thenReturn Future.successful(viewModel)
-
-        whenReady(TestAmendVariationService.compareAndUpdate(request, amlsRegistrationNumber)) {
-          updatedRequest => {
-            updatedRequest.responsiblePersons must be(convertedRequest.responsiblePersons)
-            updatedRequest must be(convertedRequest)
-          }
-        }
-      }
-
-      "user has deleted a record, added one new record and has not changed one record of responsible people" in {
-        val viewModel = DesConstants.SubscriptionViewStatusRP
-
-        val testRequest = DesConstants.amendStatusAmendVariationRP.copy(
-          businessActivities = DesConstants.testBusinessActivitiesWithDateChangeFlag.copy(
-            all = Some(DesConstants.testBusinessActivitiesAllWithDateChangeFlag.copy(
-              DateChangeFlag = Some(false)
-            ))
-          ),
-          aspOrTcsp = Some(DesConstants.testAspOrTcsp.copy(
-            supervisionDetails = Some(DesConstants.testSupervisionDetails.copy(
-              supervisorDetails = Some(DesConstants.testSupervisorDetails.copy(
-                dateChangeFlag = Some(false)
-              ))
-            ))
-          )))
-
-        when {
-          TestAmendVariationService.viewDesConnector.view(eqTo(amlsRegistrationNumber))(any())
-        } thenReturn Future.successful(viewModel)
-
-        whenReady(TestAmendVariationService.compareAndUpdate(DesConstants.updateAmendVariationRequestRP, amlsRegistrationNumber)) {
-          updatedRequest =>
-
-            updatedRequest must be(testRequest)
-        }
-      }
-    }
-
-    "successfully compare and update api6 request with api5 data" when {
-      "user has deleted a record, added a new record, modified one record and not changed one record of trading premises" in {
-        val viewModel = DesConstants.SubscriptionViewStatusTP
-
-        val testRequest = DesConstants.amendStatusAmendVariationTP.copy(
-          businessActivities = DesConstants.testBusinessActivitiesWithDateChangeFlag.copy(
-            all = Some(DesConstants.testBusinessActivitiesAllWithDateChangeFlag.copy(
-              DateChangeFlag = Some(false)
-            ))
-          ),
-          aspOrTcsp = Some(DesConstants.testAspOrTcsp.copy(
-            supervisionDetails = Some(DesConstants.testSupervisionDetails.copy(
-              supervisorDetails = Some(DesConstants.testSupervisorDetails.copy(
-                dateChangeFlag = Some(false)
-              ))
-            ))
-          ))
-        )
-
-        when {
-          TestAmendVariationService.viewDesConnector.view(eqTo(amlsRegistrationNumber))(any())
-        } thenReturn Future.successful(viewModel)
-
-        whenReady(TestAmendVariationService.compareAndUpdate(DesConstants.amendStatusDesVariationRequestTP, amlsRegistrationNumber)) {
-          updatedRequest =>
-            updatedRequest must be(testRequest)
-        }
-      }
-
-      "user has modified the previous name date of change of an RP" in {
-        val viewModel = DesConstants.SubscriptionViewStatusRP.copy(
-          responsiblePersons = DesConstants.SubscriptionViewStatusRP.responsiblePersons map {
-            rps => Seq(rps.head.copy(nameDetails = rps.head.nameDetails map {
-              nd => nd.copy(previousNameDetails = nd.previousNameDetails map {
-                pnd => pnd.copy(dateOfChange = Some("1970-01-01"))
-              })
-            }))
-          }
-        )
-        val request = DesConstants.updateAmendVariationRequestRP.copy(
-          responsiblePersons = DesConstants.updateAmendVariationRequestRP.responsiblePersons map {
-            rps => Seq(rps.tail.head)
-          }
-        )
-        val convertedRequest = DesConstants.amendStatusAmendVariationRP.copy(
-          responsiblePersons = DesConstants.amendStatusAmendVariationRP.responsiblePersons map {
-            rps => Seq(rps.tail.head.copy(
-              nameDetails = rps.tail.head.nameDetails map {
-                nd => nd.copy(previousNameDetails = nd.previousNameDetails map {
-                  pnd => pnd.copy(dateChangeFlag = Some(true))
-                })
-              }
-              , extra = rps.tail.head.extra.copy(status = Some("Updated")), dateChangeFlag = Some(false)))
-          },
-          aspOrTcsp = Some(DesConstants.testAspOrTcsp.copy(
-            supervisionDetails = Some(DesConstants.testSupervisionDetails.copy(
-              supervisorDetails = Some(DesConstants.testSupervisorDetails.copy(
-                dateChangeFlag = Some(false)
-              ))
-            ))
-          ))
-        )
-        when {
-          TestAmendVariationService.viewDesConnector.view(eqTo(amlsRegistrationNumber))(any())
-        } thenReturn Future.successful(viewModel)
-
-        whenReady(TestAmendVariationService.compareAndUpdate(request, amlsRegistrationNumber)) {
-          updatedRequest => {
-            updatedRequest.responsiblePersons must be(convertedRequest.responsiblePersons)
-            updatedRequest must be(convertedRequest)
-          }
-        }
-      }
     }
 
     "successfully compare and update api6 request with api5 1" in {
